requirements.txt
coreapi==2.1.1
coverage==4.2
dj-database-url==0.4.1
Django==1.10.4
django-cors-headers==1.3.1
django-cors-middleware==1.3.1
django-rest-swagger==2.1.0
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
gunicorn==19.6.0
itypes==1.1.0
openapi-codec==1.2.1
psycopg2==2.6.2
PyJWT==1.4.2
python-dateutil==2.6.0
python-mimeparse==1.6.0
requests==2.12.3
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
whitenoise==3.2.2
